requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
accelerate>=0.20.0

# RAG Framework
langchain>=0.0.300
langchain-community>=0.0.10
llama-index>=0.8.0

# Vector Databases
chromadb>=0.4.0
faiss-cpu>=1.7.4
pymilvus>=2.3.0

# Document Processing
pypdf>=3.15.0
python-docx>=0.8.11
markdown>=3.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.6.0

# Evaluation
rouge-score>=0.1.2
bert-score>=0.3.13

# UI and Visualization
gradio>=3.40.0
streamlit>=1.25.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Experiment Tracking
mlflow>=2.5.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.3.0

# Optional: Local LLM Support
llama-cpp-python>=0.2.0
ctransformers>=0.2.0 